requirements.txt
Django==5.0.4
psycopg2-binary>=2.9
channels==4.0.0
channels-redis==4.1.0
gunicorn>=21.2.0
uvicorn[standard]>=0.22.0
paho-mqtt==2.1.0